Type cond as u8 in ktrex_stateHandlerA02
Matches the target's byte-width clamp materialization (clrlwi r0,r0,24
at the cond join + cmplwi) instead of the int-width cmpwi.
89.46->90.76%.

+++ src/main/dll/DR/dll_0250_ktrex.c
@@ -1328,7 +1328,7 @@ int ktrex_stateHandlerA02(int obj, int runtime)
         }
         if ((int)randomGetRange(0, 0x64) <= ((u8*)p)[idx + 0x52])
         {
-            int cond;
+            u8 cond;
             u8 fe = ((KTRexArenaState*)gKTRexState)->unkFE;
             if (fe == 1)
             {
